requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-pptx==0.6.23
openai>=1.12.0
python-dotenv==1.0.1
google-cloud-aiplatform==1.42.1
pydantic==2.6.1
httpx==0.26.0
